+++ CHANGES.rst
@@ -6,11 +6,13 @@ Changelog
 
 **Added**
 
+- #1420 Allow to detach a partition from its primary sample
 - #1410 Email API
 
 
 **Changed**
 
+- #1417 Cache allowed transitions for analyses on the request
 - #1413 Improved Email Publication
 
 
@@ -20,6 +22,8 @@ Changelog
 **Fixed**
 
 - #1421 Fix Search Query for Batches Listing
+- #1418 Subscriber adapters not supported in clients listing
+- #1419 Mixed permissions for transitions in client workflow
 - #1414 Occasional "OSError: [Errno 24] Too many open files" in frontpage